+++ app/authorization/repository/repository.go
@@ -1,7 +1,3 @@
-/* Copyright © 2021. Financelime, https://financelime.com. All rights reserved.
-   Author: DmAlix. Contacts: <dmalix@financelime.com>, <dmalix@yahoo.com>
-   License: GNU General Public License v3.0, https://www.gnu.org/licenses/gpl-3.0.html */
-
 package repository
 
 import (
@@ -9,8 +5,8 @@ import (
 	"crypto/sha256"
 	"database/sql"
 	"encoding/hex"
-	"github.com/dmalix/financelime-authorization/app/authorization"
-	"github.com/dmalix/financelime-authorization/app/authorization/model"
+	"github.com/dmalix/authorization-service/app/authorization"
+	"github.com/dmalix/authorization-service/app/authorization/model"
 	"github.com/dmalix/middleware"
 	"github.com/dmalix/utils/generate"
 	"go.uber.org/zap"