+++ app/authorization/rest/router.go
@@ -1,11 +1,7 @@
-/* Copyright © 2021. Financelime, https://financelime.com. All rights reserved.
-   Author: DmAlix. Contacts: <dmalix@financelime.com>, <dmalix@yahoo.com>
-   License: GNU General Public License v3.0, https://www.gnu.org/licenses/gpl-3.0.html */
-
 package rest
 
 import (
-	"github.com/dmalix/financelime-authorization/app/authorization"
+	"github.com/dmalix/authorization-service/app/authorization"
 	"github.com/dmalix/middleware"
 	"github.com/gorilla/mux"
 	"go.uber.org/zap"
@@ -36,9 +32,9 @@ func Router(logger *zap.Logger, router *mux.Router, routerV1 *mux.Router, handle
 		handler.GetListActiveSessions(logger)).
 		Methods(http.MethodGet)
 
-	routerSession := routerV1.PathPrefix("/session/").Subrouter()
+	routerSession := routerV1.PathPrefix("/session").Subrouter()
 	routerSession.Use(middleware.Authorization(logger.Named("middlewareAuthorization")))
-	routerSession.Handle("",
+	routerSession.Handle("/",
 		handler.RevokeRefreshToken(logger)).
 		Methods(http.MethodDelete).
 		Headers(headerKeyContentType, headerValueApplicationJson)