+++ app/information/rest/rest_test.go
@@ -1,11 +1,7 @@
-/* Copyright © 2021. Financelime, https://financelime.com. All rights reserved.
-   Author: DmAlix. Contacts: <dmalix@financelime.com>, <dmalix@yahoo.com>
-   License: GNU General Public License v3.0, https://www.gnu.org/licenses/gpl-3.0.html */
-
 package rest
 
 import (
-	"github.com/dmalix/financelime-authorization/app/information/service"
+	"github.com/dmalix/authorization-service/app/information/service"
 	"github.com/dmalix/middleware"
 	"go.uber.org/zap"
 	"net/http"