+++ assets/migrate/auth.drop.sql
@@ -1,6 +1,4 @@
-/* Copyright © 2021. Financelime, https://financelime.com. All rights reserved.
-   Author: DmAlix. Contacts: <dmalix@financelime.com>, <dmalix@yahoo.com>
-   License: GNU General Public License v3.0, https://www.gnu.org/licenses/gpl-3.0.html */
+
 
 DROP TABLE IF EXISTS "public"."user";
 DROP SEQUENCE IF EXISTS "public"."user_id_seq";