+++ assets/migrate/auth.insert.sql
@@ -1,6 +1,4 @@
-/* Copyright © 2021. Financelime, https://financelime.com. All rights reserved.
-   Author: DmAlix. Contacts: <dmalix@financelime.com>, <dmalix@yahoo.com>
-   License: GNU General Public License v3.0, https://www.gnu.org/licenses/gpl-3.0.html */
+
 
 BEGIN;
 	LOCK TABLE "user", invite_code IN SHARE MODE;